# Conflicts:
#   frontend/vue.config.js

+++ backend/src/main/java/com/a506/mirinae/service/FundingService.java
@@ -15,6 +15,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -25,7 +26,7 @@ public class FundingService {
     private final FundingRepository fundingRepository;
     private final DonationRepository donationRepository;
     private final CategoryRepository categoryRepository;
-
+    
     @Transactional
     public List<FundingRes> getFundingList(String categoryName, Pageable pageable) {
         List<Funding> funding;
@@ -37,7 +38,12 @@ public class FundingService {
             funding = fundingRepository.findByCategory_Name(categoryName, PageRequest.of(pageable.getPageNumber(), pageable.getPageSize())).getContent();
         }
         for(Funding f : funding) {
-            fundingResList.add(new FundingRes(donationRepository.findDonationByFundingId(f.getId())));
+            if(f.getFundingState().equals(FundingState.ACCEPTED))
+                if(f.getDonations().size()==0)
+                    fundingResList.add(new FundingRes(f));
+                else {
+                    fundingResList.add(new FundingRes(donationRepository.findDonationByFundingId(f.getId())));
+                }
         }
         return fundingResList;
     }
@@ -68,6 +74,15 @@ public class FundingService {
                 .orElseThrow(() -> new IllegalArgumentException("해당 User가 없습니다. user ID=" + id));
         Funding funding = fundingRepository.findById(donationReq.getFundingId())
                 .orElseThrow(() -> new IllegalArgumentException("해당 펀딩이 없습니다. 펀딩 ID=" + donationReq.getFundingId()));
+        if(funding.getStartDatetime().isAfter(LocalDateTime.now()))
+            throw new IllegalArgumentException("해당 펀딩은 아직 시작되지 않았습니다!");
+        if(funding.getEndDatetime().isBefore(LocalDateTime.now()))
+            throw new IllegalArgumentException("해당 펀딩은 이미 종료되었습니다!");
+        if(!funding.getFundingState().equals(FundingState.ACCEPTED))
+            throw new IllegalArgumentException("해당 펀딩은 승인되지 않았습니다!");
+
+        donationReq.getKey();   // 지갑 비밀 키
+        
         String tx_id = "null"; //블록체인 구현 후 tx id 받기
         donationRepository.save(donationReq.toEntity(user, funding, tx_id));
     }
@@ -89,7 +104,7 @@ public class FundingService {
 
         FundingRes fundingRes = new FundingRes(donationRepository.findDonationByFundingId(funding.getId()));
 
-        return new FundingDetailRes(funding.getUser().getNickname(), fundingRes, funding.getCreatedDatetime(), funding.getStartDatetime(), funding.getEndDatetime());
+        return new FundingDetailRes(funding.getUser().getNickname(), fundingRes, funding.getCreatedDatetime(), funding.getStartDatetime(), funding.getEndDatetime(), funding.getFundingState());
     }
 
     @Transactional
@@ -98,6 +113,9 @@ public class FundingService {
             throw new IllegalArgumentException("삭제 권한이 없습니다!");
         Funding funding = fundingRepository.findById(fundingId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 펀딩이 없습니다. 펀딩 ID=" + fundingId));
+
+        if(funding.getEndDatetime().isBefore(LocalDateTime.now()))
+            throw new IllegalArgumentException("종료된 펀딩은 삭제할 수 없습니다!");
         fundingRepository.delete(funding);
     }
 }
